refactor(sandbox): move the Dockerfile layer into @kortix/shared and split it
Move the layer renderer to `@kortix/shared/sandbox` so BOTH apps/api and
apps/cli can render it (the CLI needs it for a local sandbox next), and split
it at its one real seam: staged artifacts.

- `kortixToolchainLayer(opts)` — the apt floor, the starter Python package
  floor, opencode + baked config deps, bun, the build-time warm-ups, and
  agent-browser + its Playwright Chromium. Stages nothing, so it renders
  against an EMPTY build context.
- `kortixArtifactLayer(opts)` — the contiguous tail that COPYs Kortix's own
  build outputs, unpacks/shims them, and wires ENV/WORKDIR/EXPOSE/ENTRYPOINT.
- `buildLayeredDockerfile(opts)` = user Dockerfile + toolchain + artifact.

ZERO behaviour change: the two halves are the original flat array cut in one
place, concatenated with no separator (the toolchain half carries the newline
that used to sit between them), so production output is byte-identical BY
CONSTRUCTION. Verified against the pre-split renderer across 20 render cases
(4 user Dockerfiles x 5 opt shapes) and pinned by
packages/shared/src/sandbox/__tests__/layer-split.test.ts. The rendered text is
not hashed — it reaches snapshot identity only via RUNTIME_LAYER_VERSION — so
that constant deliberately stays put.

Artifact paths stay REQUIRED: an agent-less layer would still build and still
hash to a fresh snapshot identity, failing only when a session tried to
connect. Two functions make "no artifacts" a different call, not a forgotten
field.

apps/api/src/snapshots/dockerfile-layer.ts becomes a re-export shim, so all
existing importers are untouched. The subpath export is deliberate: the root
barrel is `export *`'d into web/mobile, where SandboxTemplate/SandboxSpec would
leak and risk collisions.

Also fixes a live assertion bug in unit-dockerfile-layer.test.ts, which
asserted `import_module(mod)` while the source renders `import_module(m)` — it
could not pass against current source (that verify step was rewritten from a
heredoc to a `python3 -c` one-liner). Fixed the test, not the source.
