+++ apps/cli/package.json
@@ -24,6 +24,7 @@
     "@kortix/executor-sdk": "workspace:*",
     "@kortix/manifest-schema": "workspace:*",
     "@kortix/registry": "workspace:*",
+    "@kortix/shared": "workspace:*",
     "@kortix/starter": "workspace:*",
     "smol-toml": "^1.3.0",
     "yaml": "^2.8.2"
